requirements.txt
nose==1.3.6
pytz==2015.2
PyYAML==3.11
pymongo==3.0.1
Requests==2.9.1
